requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.104.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.343
langchain-core==0.0.7
langsmith==0.0.67
loguru==0.7.2
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==0.27.9
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
passlib==1.7.4
pinecone-client==2.2.4
pyasn1==0.5.1
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
pypdf==3.17.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.3
